W185: fifteen MATCHes, and declaration order turns out not to be dead after all.
The offset-zero census keeps paying. Fifteen more functions matched, none of
them by tuning -- every one by recovering a real stack object, a real predicate,
a real type, or a real subobject pointer.

MATCHes:
  itfMesManager 003a3530 nd5 -> 0 and 003a3680 nd5 -> 0. Both had been declared
        a "two-word signed-half register-placement floor". Checking whether the
        halfword was loaded signed showed it was, and the residual was a
        hand-expanded signed-halving correction: `maxWidth / 2` written honestly
        is the whole fix.
  itfMesManager 003a6520 nd80 -> 0. The two-state layout rebuilt as nested
        `!= 5` / `!= 4` tests so the default geometry path is inline and the
        state bodies out of line, exactly as retail.
  itfMesManager 003a67e0 nd112 -> 0. The outer predicate inverted so the
        create-and-activate path is primary.
  k_fldFrame 001acb70 nd79 -> 0, frame -0x60 -> -0x80. Distinct aligned
        line/intersection aggregates preserved, declaration order reversed to
        reproduce stack-slot order, the null-world early return recovered, and
        retail's GP-relative 1.0f used.
  k_clump 001a9760 nd104 -> 0. The typed 0xd0-byte sound-work layout
        reconstructed and an assertion filename global corrected from
        absolute-array to GP-relative scalar-address form.
  k_unit 001d38d0 nd109 -> 0. Loop locals reordered to retail liveness, the
        typed grid-value subobject pointer recovered, base/index address
        construction split.
  h_sfdply 0010d6f0 nd5 -> 0. The slot index parameter is s16, not s32 with
        repeated narrowing.
  btlVoice 002f3190 nd263 -> 0, frame -0x50 -> -0x40. An invented persistent
        boolean/result local deleted; retail branches on the call result
        directly and returns explicit 0/1.
  mt_evtMisc 0038b140 nd271 -> 0, frame -0x50 -> -0x40. A prematurely cached
        allocation counter removed, f32 pointer/copy types recovered with four
        staged float values, the absolute function-pointer slot cached across
        both frees.
  _h_camp_status 00128480 nd412 -> 0, frame -0x70 -> -0x80. The genuinely
        uninitialized parent argument recovered, the loop index made unsigned,
        and x/y kept as real float locals instead of mutating the packed
        position argument.
  datCalendar 00181580 nd258 -> 0, frame -0x40 -> -0x50. Shared result local
        and single-return layout, an invented initial clear removed, exact `== 1`
        flag predicates restored.
  fclShopMisc 004026b0 nd32 -> 0 and 0040b630 nd205 -> 0, the latter with its
        frame corrected -0x70 -> -0xa0 by rebuilding a 20-byte callback-data
        aggregate including its alignment pad.
  mdlFile 00329ba0 nd7 -> 0. See below.

**A correction to the record.** Wave 183's instrumentation work concluded that
callee-saved colouring is not reachable from source, and I passed that on as a
blanket dead end. That was too strong, and this wave proved it mid-flight:
fclShopMisc 003fea10 went nd19 -> 10 and fclCombineMisc 003d9ea0 nd13 -> 2 on
declaration order alone, and mdlFile 00329ba0 MATCHed outright by moving `depth`
ahead of `nearPlane`/`farPlane` to match retail's value-liveness order. What
wave 183 actually proved is that there is no GENERAL predictive property and
that blind permutation is useless. Ordering declarations to match the order the
values become live in RETAIL is a real lever and is now part of the method: one
measured attempt, informed by retail's prologue and first uses, then move on.
Argument-setup scheduling before a JAL remains genuinely dead; nothing has
moved it.

Near-misses created for the next pass:
  itfMesManager 003a6e30  nd478 -> 6   (ascending switch bodies for all three
                                        state sections, +0xa8 work pointer,
                                        typed entry-table pointer)
  mt_sceneFunc 003bcda0   nd163 -> 12  (two real RwMatrix stack objects and a
                                        padded 16-byte-stride vector aggregate;
                                        frame -0x80 -> -0xf0)
  btlTarget 002db480      nd260 -> 7
  btlFormation 002c70d0   nd231 -> 19
  mdlFile 00325c10        nd231 -> 6
  fclCombineMisc 003d9ea0 nd13  -> 2
  itfPanel 003ac240       nd82  -> 36
  _h_camp_status 0030c0c0 nd220 -> 108
  mt_evtMenu 0037a640     nd227 -> 90
Total normalized-diff reduction this wave: 3,781.

Also recovered along the way: the mdlFile 00335180 / 003377f0 / 00336630 family
all had their frames corrected -0xd0 -> -0xe0 by rebuilding the matrix, scaled
alpha lane, real 12-byte padding, packed colour and output bytes as one stack
aggregate. y_misc 0042cd80 went 436/576 to 568/576 with its frame corrected
-0x60 -> -0xb0 from six recovered subobject pointers.

Checkpoint: 6454 / 8211 first-party MATCH (78.6%). Zero lost, zero new
over-window addresses, zero compile errors, zero stale markers.

+++ src/Event/mt_evtMain.c
@@ -597,7 +597,7 @@ u64 FUN_00368c10(u64 param_1,u64 param_2,int param_3);
 u64 FUN_00368cb0(u64 param_1,u64 param_2,int param_3);
 u64 FUN_00368f20(u64 param_1,int param_2,u8 *param_3);
 u64 FUN_00369320(u64 param_1,int param_2,u64 param_3);
-u64 FUN_003698a0(u64 param_1,u64 param_2,int param_3);
+u32 FUN_003698a0(u64 param_1,u64 param_2,int param_3);
 u64 FUN_00369a20(int param_1,u64 param_2,int param_3);
 u64 FUN_0036be10(u64 param_1,u64 param_2,u32 param_3);
 u64 FUN_0036cc30(u64 param_1,u64 param_2,u8 *param_3);
@@ -1044,6 +1044,7 @@ void FUN_00361270(int param_1)
   int iVar2;
 
   u16 *extraout_a1_lo;
+
 
   u16 *puVar3;
 
@@ -10116,7 +10117,7 @@ u64 FUN_00369320(u64 param_1,int param_2,u64 param_3)
 // FUN_003698A0 NONMATCHING
 
 
-u64 FUN_003698a0(u64 param_1,u64 param_2,int param_3)
+u32 FUN_003698a0(u64 param_1,u64 param_2,int param_3)
 
 
 
@@ -10132,7 +10133,7 @@ u64 FUN_003698a0(u64 param_1,u64 param_2,int param_3)
 
   int iVar5;
 
-  u64 extraout_t0;
+  int extraout_t0;
